requirements.txt
discord.py==2.3.1
requests==2.31.0
api-python-client
google-auth google-auth-oauthlib google-auth-httplib2
openai
forex_python